refactor: Move InkResponse to standalone file as share interface
